assignments/assignment3/requirements.txt
matplotlib==3.8.4
numpy==1.26.4
pandas==2.2.2
Pillow==10.3.0
scikit_learn==1.4.2
tensorflow_cpu==2.16.1
tqdm==4.66.2
